+++ packages/nimble-components/src/table-column/anchor/tests/table-column-anchor.spec.ts
@@ -57,6 +57,16 @@ describe('TableColumnAnchor', () => {
         await disconnect();
     });
 
+    it('should export its tag', () => {
+        expect(tableColumnAnchorTag).toBe('nimble-table-column-anchor');
+    });
+
+    it('can construct an element instance', () => {
+        expect(
+            document.createElement('nimble-table-column-anchor')
+        ).toBeInstanceOf(TableColumnAnchor);
+    });
+
     it('reports column configuration valid', async () => {
         await connect();
         await waitForUpdatesAsync();